requirements.txt
# Core
streamlit==1.20.0
requests==2.32.3
ollama==0.3.0
tensorflow==2.10.1

# Data handling
pandas==1.5.3
numpy==1.26.4
openpyxl==3.1.5
xlsxwriter==3.2.9

# Document / file manipulation
python-docx==1.2.0
pdfplumber==0.11.7
PyPDF2==3.0.1
python-pptx==1.0.2
pillow==10.4.0
lxml==5.3.0
et_xmlfile==2.0.0
pypdfium2==4.30.0

# Validation and schema
attrs==24.2.0
typing_extensions==4.12.2
jsonschema==4.25.0
jsonschema-specifications==2024.10.1
referencing==0.32.0
rpds-py==0.13.2
validators==0.35.0

# Visualization and UI
altair==4.2.2
pydeck==0.9.1
matplotlib==3.9.2
protobuf==3.19.6
blinker==1.9.0
watchdog==4.0.1
tornado==6.4.1
colorama==0.4.6
click==8.1.7
tenacity==8.3.0
rich==13.8.0
markdown-it-py==3.0.0
mdurl==0.1.2
toml==0.10.2
Jinja2==3.1.4
MarkupSafe==3.0.2
tabulate==0.9.0

# Database
sqlalchemy==2.0.35

# Networking
httpx==0.27.0
httpcore==1.0.5
h11==0.14.0
urllib3==2.2.3
certifi==2024.8.30
